Initial commit
This is the initial commit before ripping apart the frame-making code.
Currently that code resides in each plugin that wants it, which makes
the test harnesses more simple, but means that we're extracting frames
once for the gif-maker and a second time for the keyword tagger. Need to
fix this, so this is a checkpoint commit before that refactor.